Rename utilities_retrieval.py to retrieval.py
