+++ financial_assistant/src/tools_yahoo_news.py
@@ -11,9 +11,9 @@ from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain_core.pydantic_v1 import BaseModel, Field
 from langchain_core.tools import tool
 
+from financial_assistant.src.retrieval import get_qa_response
 from financial_assistant.src.tools import coerce_str_to_list, get_logger
 from financial_assistant.src.tools_stocks import retrieve_symbol_list
-from financial_assistant.src.utilities_retrieval import get_qa_response
 from financial_assistant.streamlit.constants import *
 
 RETRIEVE_HEADLINES = False
